grains: updated tests to v1.2.0 (#1511)
Closes #1504

Note that no changes were made to the tests since Python is in-line with the [canonical data](https://github.com/exercism/problem-specifications/blob/master/exercises/grains/canonical-data.json). The canonical-data was updated to have error objects instead of -1 when an error would occur, and the Python tests are already in-line with that policy.