[book-store] Simplify example.py (#2175)
The existing example.py is quite complex and not particulary efficient,
even for quite small lists of books. This example.py is much simpler,
but could be made considerably more efficient through a bit of a caching
and some sensible short-circuit returns. These are left to the student
to figure out, however.