[Leap]: Add `calendar-isleap` approach and improved benchmarks (#3583)
* [Leap]: Add `calendar-isleap` approach and improved benchmarks

* added comments on `isleap()` implementation and operator precedence

* Approach and Chart Edits

Various language edits and touchups.

Added chart reference and alt-textm as well as link to long description.  Probably unnecessary for this particular article but wanted to establish an example of what to do when alt-texting a chart.  See https://www.w3.org/WAI/tutorials/images/complex/ for more information and examples on how to alt-text complex images.

* Fixed Awkward Chart Intro

* Another try on alt-text

* Trying Yet Again

* Attempt to align figcaption

* Giving up and going for description location in alt

---------

Co-authored-by: BethanyG <BethanyG@users.noreply.github.com>